+++ strato/tools/blockapps-tools/src/BlockApps/Tools/Code.hs
@@ -2,6 +2,7 @@ module BlockApps.Tools.Code where
 
 import BlockApps.Tools.DumpLevelDB
 import Blockchain.Strato.Model.Code
+import Data.ByteString.Short (toShort)
 
 doit :: String -> String -> IO ()
-doit dbtype h = showKeyVal (show . Code) dbtype "code" (if h == "-" then Nothing else Just h)
+doit dbtype h = showKeyVal (show . Code . toShort) dbtype "code" (if h == "-" then Nothing else Just h)
